requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aerich==0.9.2
    # via -r backend/requirements.in
aiosqlite==0.22.1
    # via tortoise-orm
anchorpy==0.21.0
    # via -r requirements.in
anchorpy-core==0.2.0
    # via anchorpy
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   aerich
    #   httpx
    #   starlette
    #   tortoise-orm
    #   watchfiles
asyncclick==8.3.0.7
    # via aerich
asyncpg==0.31.0
    # via tortoise-orm
attrs==25.4.0
    # via sumtypes
base58==2.1.1
    # via -r backend/requirements.in
based58==0.1.1
    # via anchorpy
bitarray==3.8.0
    # via eth-account
borsh-construct==0.1.0
    # via anchorpy
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
cfgv==3.5.0
    # via pre-commit
ckzg==2.1.5
    # via eth-account
click==8.3.1
    # via uvicorn
construct==2.10.68
    # via construct-typing
construct-typing==0.5.6
    # via
    #   anchorpy
    #   borsh-construct
    #   solana
cytoolz==1.1.0
    # via eth-utils
dictdiffer==0.9.0
    # via aerich
distlib==0.4.0
    # via virtualenv
eth-abi==6.0.0b1
    # via eth-account
eth-account==0.13.7
    # via -r backend/requirements.in
eth-hash==0.7.1
    # via eth-utils
eth-keyfile==0.8.1
    # via eth-account
eth-keys==0.7.0
    # via
    #   eth-account
    #   eth-keyfile
eth-rlp==2.2.0
    # via eth-account
eth-typing==5.2.1
    # via
    #   eth-abi
    #   eth-keys
    #   eth-utils
eth-utils==5.3.1
    # via
    #   eth-abi
    #   eth-account
    #   eth-keyfile
    #   eth-keys
    #   eth-rlp
    #   rlp
fastapi==0.129.0
    # via -r backend/requirements.in
filelock==3.24.3
    # via
    #   python-discovery
    #   virtualenv
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hexbytes==1.3.1
    # via
    #   eth-account
    #   eth-rlp
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   -r backend/requirements.in
    #   solana
identify==2.6.16
    # via pre-commit
idna==3.11
    # via
    #   anyio
    #   httpx
iso8601==2.1.0
    # via tortoise-orm
jsonalias==0.1.1
    # via
    #   anchorpy-core
    #   solders
nodeenv==1.10.0
    # via pre-commit
parsimonious==0.10.0
    # via eth-abi
platformdirs==4.9.2
    # via
    #   python-discovery
    #   virtualenv
pre-commit==4.5.1
    # via -r requirements.in
pycryptodome==3.23.0
    # via eth-keyfile
pydantic==2.12.5
    # via
    #   -r backend/requirements.in
    #   eth-account
    #   eth-utils
    #   fastapi
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via -r backend/requirements.in
pyheck==0.1.5
    # via anchorpy
pypika-tortoise==0.6.3
    # via tortoise-orm
python-discovery==1.1.0
    # via virtualenv
python-dotenv==1.2.1
    # via
    #   -r backend/requirements.in
    #   pydantic-settings
    #   uvicorn
pytz==2025.2
    # via tortoise-orm
pyyaml==6.0.3
    # via
    #   pre-commit
    #   uvicorn
regex==2026.2.19
    # via parsimonious
rlp==4.1.0
    # via
    #   eth-account
    #   eth-rlp
solana==0.36.6
    # via
    #   -r backend/requirements.in
    #   anchorpy
solders==0.26.0
    # via
    #   anchorpy
    #   solana
starlette==0.52.1
    # via fastapi
sumtypes==0.1a6
    # via borsh-construct
toml==0.10.2
    # via anchorpy
toolz==0.11.2
    # via
    #   anchorpy
    #   cytoolz
tortoise-orm[asyncpg]==0.25.4
    # via
    #   -r backend/requirements.in
    #   aerich
typing-extensions==4.15.0
    # via
    #   anyio
    #   eth-typing
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   solana
    #   solders
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-settings
uvicorn[standard]==0.40.0
    # via -r backend/requirements.in
uvloop==0.22.1
    # via uvicorn
virtualenv==21.1.0
    # via pre-commit
watchfiles==1.1.1
    # via uvicorn
websockets==15.0
    # via
    #   solana
    #   uvicorn
